refactor(rust-api): extract model catalog/jobs/helpers into models.rs (sc-1639 Phase 1, slice 2)
Second behavior-preserving slice of the sc-1639 lib.rs Value-cluster split.
Moves the model-exclusive code out of lib.rs into a new models.rs (1,080 lines):
- handlers: list_models, create_model_download_job, create_model_convert_job,
  delete_model, create_model_import_job + import plumbing
- catalog/download-size: model_catalog, model_download(+context/supported),
  manifest_download_size_bytes, estimate_huggingface_download_size(+uncached),
  download_size_from_siblings, ModelSizeCache, DownloadContext
- manifest resolve: resolve_model_manifest_entry, merge_model_manifest_entry
- install/artifact/status: model_is_installed, model_artifact_paths,
  model_manifest_installed_path, MlxCatalogStatus, mlx_catalog_status

Shared HF-cache/download/pattern/format/artifact helpers (used by loras +
training too) stay at root; models.rs reaches them via `use super::*`. Same
sc-1636 pattern: pub(crate) on the external surface + `use models::*` re-export
so lib.rs/generation.rs/training.rs/tests.rs call-sites resolve unchanged.
MlxCatalogStatus fields and ModelSizeCache get/insert are pub(crate) because
tests.rs (now a sibling module) reads them.

lib.rs 5,017 -> 3,945. cargo test (62 rust-api), npm run rust:check, and
cargo build --features embed-web all green.
